reload units after linking credentials

+++ meta-printnanny/recipes-core/octoprint-apps/files/octoprint-link-confd.sh
@@ -12,6 +12,19 @@ OCTOPRINT_CONFD="/run/systemd/system/octoprint-env.service.d"
 mkdir -p "$JANUS_CONFD"
 mkdir -p "$OCTOPRINT_CONFD"
 
+CHANGED="0"
+
 # symlink janus admin and api tokens so printnanny can initialize webrtc sessions (via octoprint plugin)
-ln -s "${JANUS_CONFD}/${JANUS_ADMIN_FILE}" "${OCTOPRINT_CONFD}/${JANUS_ADMIN_FILE}"
-ln -s "${JANUS_CONFD}/${JANUS_TOKEN_FILE}" "${OCTOPRINT_CONFD}/${JANUS_TOKEN_FILE}"
+if [ -f "${OCTOPRINT_CONFD}/${JANUS_ADMIN_FILE}" ]; then
+    ln -s "${JANUS_CONFD}/${JANUS_ADMIN_FILE}" "${OCTOPRINT_CONFD}/${JANUS_ADMIN_FILE}"
+    CHANGED="1"
+fi
+
+if [ -f "${JANUS_CONFD}/${JANUS_TOKEN_FILE}" ]; then
+    ln -s "${JANUS_CONFD}/${JANUS_ADMIN_FILE}" "${JANUS_CONFD}/${JANUS_TOKEN_FILE}"
+    CHANGED="1"
+fi
+
+if [ "$CHANGED" == "1" ]; then
+    systemctl daemon-reload
+fi
